- install package for routing - resture app moving content to pages - add purchase, sell and no match pages - add route funcionality on succesful sign up/sign in
