+++ client/package.json
@@ -12,6 +12,7 @@
     "react-dom": "^17.0.1",
     "react-icons": "^4.2.0",
     "react-places-autocomplete": "^7.3.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.2",
     "web-vitals": "^1.1.0"
   },
