+++ client/src/components/NoMatch.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const NoMatch = () => <h1>Nothing Here... Move!!!</h1>;
+
+export default NoMatch;
