+++ client/src/components/Signin.js
@@ -1,7 +1,8 @@
 import React, { useState } from "react";
+import { useHistory } from "react-router-dom";
 
 import axiosApi from "../api/axiosApi";
-import "./AuthForm.css";
+import "./AuthPage.css";
 import { saveToken } from "../utils";
 
 const Signin = () => {
@@ -11,6 +12,7 @@ const Signin = () => {
   };
 
   const [values, setValues] = useState(defaultValues);
+  const history = useHistory();
 
   const handleChange = event => {
     setValues({ ...values, [event.target.name]: event.target.value });
@@ -32,6 +34,7 @@ const Signin = () => {
     };
 
     saveToken(userDetails);
+    history.push("/buy");
   };
 
   return (
